requirements.txt
chainlit>=2.5.5
pydantic>=2.10.4,<2.11.0
python-dotenv>=1.0.0
asyncpg>=0.30.0
boto3>=1.38.24
pydantic-ai-slim[openai]>=0.2.14
